+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 
+# 0.3.0
+
+- The create and drop have been renamed to `create_table` and `drop_table` respectively.
+- Add a shortcut `SomeModel.object.create(**state)` method
+- Allow passing a db connection to manager methods (
+    get, get_or_create, filter, delete, etc...) to better support transactions
+
+
+# 0.2.4
+
+- Fix the nosql serialization registry not being loaded properly
+
+# 0.2.3
+
+- Fix packaging issue with 0.2.2
+
 # 0.2.2
 
 - Fix bug with fk types